Clickable Pivots, more LaTeX and Focus
The pivots are clickable after clicking the Select Pivot button, and
doing so auto fills the input box and highlights the selection.  All the
string filtering for Matrix creation was moved to the constructor, so
both pages can enter matrices in LaTeX mode.  More focus changes to the
Matrix Operations page.